changed response type to include a status
